+++ src/Check/Memcached.php
@@ -1,22 +1,22 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-diagnostics for the canonical source repository
- * @copyright Copyright (c) 2013-2018 Zend Technologies USA Inc. (https://www.zend.com)
+ * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-diagnostics/blob/master/LICENSE.md New BSD License
  */
 
 namespace ZendDiagnostics\Check;
 
 use Exception;
-use Memcache as MemcacheService;
+use Memcached as MemcachedService;
 use InvalidArgumentException;
 use ZendDiagnostics\Result\Failure;
 use ZendDiagnostics\Result\Success;
 
 /**
- * Check if MemCache extension is loaded and given server is reachable.
+ * Check if MemCached extension is loaded and given server is reachable.
  */
-class Memcache extends AbstractCheck
+class Memcached extends AbstractCheck
 {
     /**
      * @var string
@@ -31,7 +31,8 @@ class Memcache extends AbstractCheck
     /**
      * @param string $host
      * @param int    $port
-     * @throws InvalidArgumentException
+     * @throws InvalidArgumentException if host is not a string value
+     * @throws InvalidArgumentException if port is less than 1
      */
     public function __construct($host = '127.0.0.1', $port = 11211)
     {
@@ -59,14 +60,14 @@ class Memcache extends AbstractCheck
      */
     public function check()
     {
-        if (! class_exists('Memcache', false)) {
-            return new Failure('Memcache extension is not loaded');
+        if (! class_exists('Memcached', false)) {
+            return new Failure('Memcached extension is not loaded');
         }
 
         try {
-            $memcache = new MemcacheService();
-            $memcache->addServer($this->host, $this->port);
-            $stats = @$memcache->getExtendedStats();
+            $memcached = new MemcachedService();
+            $memcached->addServer($this->host, $this->port);
+            $stats = @$memcached->getStats();
 
             $authority = sprintf('%s:%d', $this->host, $this->port);
 
@@ -76,9 +77,9 @@ class Memcache extends AbstractCheck
                 || false === $stats[$authority]
             ) {
                 // Attempt a connection to make sure that the server is really down
-                if (! @$memcache->connect($this->host, $this->port)) {
+                if (! @$memcached->getLastDisconnectedServer($this->host, $this->port)) {
                     return new Failure(sprintf(
-                        'No memcache server running at host %s on port %s',
+                        'No memcached server running at host %s on port %s',
                         $this->host,
                         $this->port
                     ));
@@ -89,7 +90,7 @@ class Memcache extends AbstractCheck
         }
 
         return new Success(sprintf(
-            'Memcache server running at host %s on port %s',
+            'Memcached server running at host %s on port %s',
             $this->host,
             $this->port
         ));